fix: restore settings coverage gate
